Fix server rsp issue
There is a known issue in the destination API server where, when requesting some specific data, the server may close the connection on its own before finishing receiving the response.
To address this provisionally, if Axios throws an `maxContentLength size of -1 exceeded` error, it will ignore the error and continue.